3.6 delta test

+++ kodilla-basic-test/src/main/java/com/kodilla/basic_assertion/Application.java
@@ -21,7 +21,7 @@ public class Application {
             System.out.println("Metoda subtract nie działa poprawnie dla liczb " + a + " i " + b);
         }
 
-        int squaredResult = calculator.squared(a);
+        int squaredResult = (int) calculator.squared(a);
         boolean correct3 = ResultChecker.assertEquals(25, squaredResult);
         if (correct3) {
             System.out.println("Metoda squared działa poprawnie dla liczby " + a);
